new function + example how to use EJSCREEN API as input to batch.summarizer
The function I added, prep.bufferapi.for.summarizer(), has an example script showing how to use that script to get batch buffer results (raw output table for a set of circular buffers) from the EJSCREEN API, save that output as a csv in the right format, ready to be imported to the batch.summarizer R Shiny app.  At least that provides a slow and manual process for getting batch buffer results and an interactive summary of them.